# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AMAZING SOULS API",
        "description": "API documentation for AMAZING SOULS API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications for the authenticated user",
                "operationId": "listNotifications",
                "responses": {
                    "200": {
                        "description": "Notifications fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/notifications/test": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a test notification",
                "operationId": "testNotification",
                "responses": {
                    "200": {
                        "description": "Test notification created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/uploads/{media}": {
            "delete": {
                "tags": [
                    "Uploads"
                ],
                "summary": "Delete media by id",
                "operationId": "deleteMediaById",
                "parameters": [
                    {
                        "name": "media",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/uploads": {
            "delete": {
                "tags": [
                    "Uploads"
                ],
                "summary": "Delete media by path or URL",
                "operationId": "deleteMediaByPath",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "path"
                                ],
                                "properties": {
                                    "path": {
                                        "type": "string",
                                        "example": "uploads/originals/2026/03/15/file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Sanctum personal access token. Swagger will send it as Authorization: Bearer {token}.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}